pyproject.toml
[tool.poetry]
name = "pylogger-slack"
version = "0.0.1"
description = "Python logger with an option to format it for YAML/JSON format and support for Slack Notification. (Helpful for monitoring codebase in production/testing)"
authors = ["Binay Kumar Pradhan <bpradhan.asli@gmail.com>"]
#readme = "README.md"
license = "GPL-3.0"
repository = "https://github.com/VinayIN/pylogger-slack.git"
keywords = ["logger", "slack notification", "logger yaml", "logger json", "python logging"]
packages = [
    { include = "pylogger_slack" }
]
include = ["logger.yml"]

[tool.poetry.dependencies]
python = "^3.11"
ecs-logging = "2.0.0"
PyYAML = "6.0"
slack-sdk = "3.18.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"